refactor(frontend): Rename component OpenCryptoPayWizard
